extra guard against unlinking twice, https://github.com/phetsims/molecules-and-light/issues/151

+++ js/moleculesandlight/view/ObservationWindow.js
@@ -161,7 +161,7 @@ define( function( require ) {
 
       photonAbsorptionModel.photons.addItemRemovedListener( function removalListener( removedPhoton ) {
         if ( removedPhoton === addedPhoton ) {
-          addedPhoton.locationProperty.unlink( photonPositionObserver );
+          addedPhoton.locationProperty.hasListener( photonPositionObserver ) && addedPhoton.locationProperty.unlink( photonPositionObserver );
           photonLayer.removeChild( photonNode );
           photonAbsorptionModel.photons.removeItemRemovedListener( removalListener );
         }
